recommendation_system/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
httpx==0.24.0
pydantic==1.10.7
numpy==1.24.3
torch==2.0.0
transformers==4.36.0
accelerate==0.25.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4
requests>=2.31.0
alembic==1.11.1
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
watchdog==3.0.0